figured out module layout, added definitions for buffer page etc., added parse_func to bparse (for deserialization without returning subsequent)

+++ src/dbsystem/tx/tx.go
@@ -6,7 +6,7 @@ type Tx struct {
 	Status Status
 }
 
-type Id = uint64
+type Id = uint32
 
 type IsolationLevel = uint8
